fix(googlechat): preserve reply text after typing update failures
