+++ extensions/googlechat/src/monitor.ts
@@ -1,7 +1,3 @@
-import {
-  deliverTextOrMediaReply,
-  resolveSendableOutboundReplyParts,
-} from "openclaw/plugin-sdk/reply-payload";
 import { normalizeOptionalLowercaseString } from "openclaw/plugin-sdk/text-runtime";
 import type { OpenClawConfig } from "../runtime-api.js";
 import {
@@ -10,15 +6,10 @@ import {
   resolveWebhookPath,
 } from "../runtime-api.js";
 import { type ResolvedGoogleChatAccount } from "./accounts.js";
-import {
-  downloadGoogleChatMedia,
-  deleteGoogleChatMessage,
-  sendGoogleChatMessage,
-  uploadGoogleChatAttachment,
-  updateGoogleChatMessage,
-} from "./api.js";
+import { downloadGoogleChatMedia, sendGoogleChatMessage } from "./api.js";
 import { type GoogleChatAudienceType } from "./auth.js";
 import { applyGoogleChatInboundAccessPolicy, isSenderAllowed } from "./monitor-access.js";
+import { deliverGoogleChatReply } from "./monitor-reply-delivery.js";
 import {
   handleGoogleChatWebhookRequest,
   registerGoogleChatWebhookTarget,
@@ -335,152 +326,6 @@ async function downloadAttachment(
     attachment.contentName,
   );
   return { path: saved.path, contentType: saved.contentType };
-}
-
-async function deliverGoogleChatReply(params: {
-  payload: { text?: string; mediaUrls?: string[]; mediaUrl?: string; replyToId?: string };
-  account: ResolvedGoogleChatAccount;
-  spaceId: string;
-  runtime: GoogleChatRuntimeEnv;
-  core: GoogleChatCoreRuntime;
-  config: OpenClawConfig;
-  statusSink?: (patch: { lastInboundAt?: number; lastOutboundAt?: number }) => void;
-  typingMessageName?: string;
-}): Promise<void> {
-  const { payload, account, spaceId, runtime, core, config, statusSink } = params;
-  // Use let so we can clear it after a successful delete — if we leave it set, sendText
-  // would try to updateGoogleChatMessage on an already-deleted message and silently drop
-  // all text content (the update error is caught before firstTextChunk flips to false).
-  let typingMessageName = params.typingMessageName;
-  const reply = resolveSendableOutboundReplyParts(payload);
-  const mediaCount = reply.mediaCount;
-  const hasMedia = reply.hasMedia;
-  const text = reply.text;
-  let firstTextChunk = true;
-  let suppressCaption = false;
-
-  if (hasMedia) {
-    if (typingMessageName) {
-      try {
-        await deleteGoogleChatMessage({
-          account,
-          messageName: typingMessageName,
-        });
-        // Clear after successful delete so the sendText path below does not attempt to
-        // update a message that no longer exists.
-        typingMessageName = undefined;
-      } catch (err) {
-        runtime.error?.(`Google Chat typing cleanup failed: ${String(err)}`);
-        // typingMessageName is still set here: the delete threw before the reset ran.
-        // The guard below satisfies TypeScript's conservative catch-block narrowing.
-        if (typingMessageName) {
-          const fallbackText = reply.hasText
-            ? text
-            : mediaCount > 1
-              ? "Sent attachments."
-              : "Sent attachment.";
-          try {
-            await updateGoogleChatMessage({
-              account,
-              messageName: typingMessageName,
-              text: fallbackText,
-            });
-            suppressCaption = Boolean(text.trim());
-          } catch (updateErr) {
-            runtime.error?.(`Google Chat typing update failed: ${String(updateErr)}`);
-            // Also clear here so sendText falls through to sendGoogleChatMessage
-            // instead of repeatedly retrying the unavailable message and dropping text.
-            typingMessageName = undefined;
-          }
-        }
-      }
-    }
-  }
-
-  const chunkLimit = account.config.textChunkLimit ?? 4000;
-  const chunkMode = core.channel.text.resolveChunkMode(config, "googlechat", account.accountId);
-  await deliverTextOrMediaReply({
-    payload,
-    text: suppressCaption ? "" : reply.text,
-    chunkText: (value) => core.channel.text.chunkMarkdownTextWithMode(value, chunkLimit, chunkMode),
-    sendText: async (chunk) => {
-      try {
-        if (firstTextChunk && typingMessageName) {
-          await updateGoogleChatMessage({
-            account,
-            messageName: typingMessageName,
-            text: chunk,
-          });
-        } else {
-          await sendGoogleChatMessage({
-            account,
-            space: spaceId,
-            text: chunk,
-            thread: payload.replyToId,
-          });
-        }
-        firstTextChunk = false;
-        statusSink?.({ lastOutboundAt: Date.now() });
-      } catch (err) {
-        runtime.error?.(`Google Chat message send failed: ${String(err)}`);
-        if (firstTextChunk && typingMessageName) {
-          // The updateGoogleChatMessage call failed. Clear both flags so
-          // subsequent chunks fall through to sendGoogleChatMessage instead of
-          // retrying the unavailable typing-indicator message and silently
-          // dropping all remaining text.
-          typingMessageName = undefined;
-          firstTextChunk = false;
-        }
-      }
-    },
-    sendMedia: async ({ mediaUrl, caption }) => {
-      try {
-        const loaded = await core.channel.media.fetchRemoteMedia({
-          url: mediaUrl,
-          maxBytes: (account.config.mediaMaxMb ?? 20) * 1024 * 1024,
-        });
-        const upload = await uploadAttachmentForReply({
-          account,
-          spaceId,
-          buffer: loaded.buffer,
-          contentType: loaded.contentType,
-          filename: loaded.fileName ?? "attachment",
-        });
-        if (!upload.attachmentUploadToken) {
-          throw new Error("missing attachment upload token");
-        }
-        await sendGoogleChatMessage({
-          account,
-          space: spaceId,
-          text: caption,
-          thread: payload.replyToId,
-          attachments: [
-            { attachmentUploadToken: upload.attachmentUploadToken, contentName: loaded.fileName },
-          ],
-        });
-        statusSink?.({ lastOutboundAt: Date.now() });
-      } catch (err) {
-        runtime.error?.(`Google Chat attachment send failed: ${String(err)}`);
-      }
-    },
-  });
-}
-
-async function uploadAttachmentForReply(params: {
-  account: ResolvedGoogleChatAccount;
-  spaceId: string;
-  buffer: Buffer;
-  contentType?: string;
-  filename: string;
-}) {
-  const { account, spaceId, buffer, contentType, filename } = params;
-  return await uploadGoogleChatAttachment({
-    account,
-    space: spaceId,
-    filename,
-    buffer,
-    contentType,
-  });
 }
 
 export function monitorGoogleChatProvider(options: GoogleChatMonitorOptions): () => void {
